slight refactor to improve naming scheme
